requirements.txt
streamlit
Pillow
pytesseract
torch
transformers
